services/editor/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sidekick-editor"
version = "0.1.0"
description = "Editor agent — draft tier (Phase 5)"
requires-python = ">=3.12"
dependencies = [
    "sidekick-core",
    "typer>=0.24.1",
    "python-dotenv>=1.0",
    "deepagents>=0.4.11",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
]

[project.scripts]
sidekick-editor = "sidekick.editor.cli:main"

[tool.uv.sources]
sidekick-core = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/sidekick"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
